Move cli argument parsing to cli.ts
