+++ amp-toolbox-linter/package.json
@@ -10,7 +10,7 @@
     "amp-toolbox-cache-url": "^1.0.6",
     "amphtml-validator": "^1.0.23",
     "cheerio": "^1.0.0-rc.2",
-    "commander": "^2.19.0",
+    "commander": "^2.20.0",
     "debug": "^4.1.1",
     "escape-html": "^1.0.3",
     "execa": "^1.0.0",
@@ -26,7 +26,7 @@
     "build": "tsc --module commonjs --target es2018 --esModuleInterop --noImplicitAny --strictNullChecks src/*.ts tests/*.ts",
     "test": "cd tests && for t in helper.js local.js network.js ; do ( test -e $t || $npm_execpath run build ) && ( node $t | tap-parser -t ) || exit 1 ; done ; exit 0",
     "lint": "prettier --check 'src/**/*.ts' 'tests/**/*.ts'",
-    "watch": "ls tsconfig.json tslint.json src/*.ts tests/*.ts | entr -c -s 'clear ; $npm_execpath run build ; date'",
+    "watch": "tsc --watch --module commonjs --target es2018 --esModuleInterop --noImplicitAny --strictNullChecks src/*.ts tests/*.ts",
     "package": "$npm_execpath run lint && pack build",
     "publish": "pack publish --any-branch",
     "cache": "cd src && curl -sS --remote-name-all https://cdn.ampproject.org/v0/validator.js && curl -sS --remote-name-all https://cdn.ampproject.org/caches.json"
